+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -2,16 +2,22 @@ package roomescape.service;
 
 import java.util.List;
 import org.springframework.stereotype.Service;
+import roomescape.common.exception.ConflictException;
+import roomescape.common.exception.NotFoundException;
+import roomescape.dao.ReservationDao;
 import roomescape.dao.ReservationTimeDao;
-import roomescape.domain.ReservationTime;
+import roomescape.domain.reservation.time.ReservationTime;
+import roomescape.dto.request.ReservationTimeRequest;
 import roomescape.dto.response.ReservationTimeResponse;
 
 @Service
 public class ReservationTimeService {
     private final ReservationTimeDao reservationTimeDao;
+    private final ReservationDao reservationDao;
 
-    public ReservationTimeService(ReservationTimeDao reservationTimeDao) {
+    public ReservationTimeService(ReservationTimeDao reservationTimeDao, ReservationDao reservationDao) {
         this.reservationTimeDao = reservationTimeDao;
+        this.reservationDao = reservationDao;
     }
 
     public List<ReservationTimeResponse> findAll() {
@@ -22,13 +28,25 @@ public class ReservationTimeService {
                 .toList();
     }
 
-    public ReservationTimeResponse save(ReservationTime reservationTime) {
-        ReservationTime time = reservationTimeDao.save(reservationTime);
+    public ReservationTimeResponse save(ReservationTimeRequest request) {
+        ReservationTime time = new ReservationTime(
+                request.startAt()
+        );
 
-        return ReservationTimeResponse.from(time);
+        ReservationTime saved = reservationTimeDao.save(time);
+
+        return ReservationTimeResponse.from(saved);
     }
 
     public void delete(Long id) {
+        if (!reservationTimeDao.existsById(id)) {
+            throw new NotFoundException("존재하지 않는 시간입니다.");
+        }
+
+        if (reservationDao.existsByTimeId(id)) {
+            throw new ConflictException("예약이 존재하는 시간은 삭제할 수 없습니다.");
+        }
+
         reservationTimeDao.delete(id);
     }
 }
